Arizona/pld_agent/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "arizona"
version = "2.0.1"
description = "Arizona — Agente PLD/AML + Oficial de Cumplimiento PLD/FT"
requires-python = ">=3.11"
dependencies = [
    "asyncpg>=0.29",
    "fastapi>=0.115",
    "uvicorn>=0.34",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "openai>=1.12",
    "numpy>=1.26",
    "httpx>=0.27",
    "azure-search-documents>=11.6",
    "pyodbc>=5.1",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.24"]

[tool.setuptools.packages.find]
include = ["pld_agent*"]
